daily_briefing/requirements.txt
fastapi
uvicorn
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
playwright
beautifulsoup4
python-dotenv
google-genai
feedparser
requests
tenacity
